# src/knowledge_base/api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Unified Knowledge Base System API
  description: |
    # Unified Knowledge Base System API
    
    This API provides access to the Unified Knowledge Base System, allowing you to:
    
    - Add, update, and retrieve documents
    - Query the knowledge base with natural language
    - Manage users and API keys
    - Monitor system health and performance
    - Perform administrative tasks
    
    ## Authentication
    
    Most endpoints require authentication using either:
    - API key (via `X-API-Key` header)
    - Bearer token (via `Authorization` header)
    
    ## WebSocket Support
    
    Real-time communication is available through WebSocket at `/ws/{client_id}`.
    
    ## Rate Limiting
    
    API requests are subject to rate limiting based on your account tier.
  version: 1.0.0
  termsOfService: https://example.com/terms/
  contact:
    name: API Support
    url: https://example.com/support
    email: support@example.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: https://api.example.com/v1
    description: Production server
  - url: https://staging-api.example.com/v1
    description: Staging server
  - url: http://localhost:8000
    description: Local development server
...
paths:
...
  /ws/{client_id}:
    get:
      summary: WebSocket endpoint
      description: |
        Connect to the WebSocket API for real-time communication.
        
        ## Message Format
        
        All messages follow this format:
        
        ```json
        {
          "request_id": "unique-request-id",
          "request_type": "query",
          "payload": {
            // Request-specific payload
          }
        }
        ```
        
        ## Available Operations
        
        - Query: `request_type: "query"`
        - Subscribe: `request_type: "subscribe"`
        - Unsubscribe: `request_type: "unsubscribe"`
        
        See the WebSocket documentation for more details.
      operationId: websocketEndpoint
      tags:
        - WebSocket
      parameters:
        - name: client_id
          in: path
          description: Client identifier
          required: true
          schema:
            type: string
      responses:
        '101':
          description: WebSocket connection established
components:
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key authentication
    
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token authentication
security:
  - ApiKeyAuth: []
  - BearerAuth: []